Convert the queue status badge off Redux
First of three PRs for Queue. The slice has independent status, details
and paged sub-sections; this takes status only, so the sidebar badge and
its SignalR path convert without touching the Queue page.

`useQueueStatus` replaces `createQueueStatusSelector`, and the `queue/status`
SignalR handler writes to the query cache with `setQueryData` rather than
dispatching a redux update -- the last `update()` dispatch in
SignalRListener, so that import goes too.

`QueueStatus` used to refetch itself when SignalR reconnected, because the
badge misses every message while the connection is down. That refresh moves
into `handleReconnected` alongside the other reconnect invalidations.

Eros keeps `includeUnknownMovieItems`, which Sonarr replaced with a queue
filter. Their hook always reports totalCount; ours still derives the badge
from the option, which stays in redux until the paged section converts.
