+++ frontend/src/App/State/QueueAppState.ts
@@ -1,21 +1,10 @@
 import Queue from 'typings/Queue';
 import AppSectionState, {
   AppSectionFilterState,
-  AppSectionItemState,
   Error,
   PagedAppSectionState,
   TableAppSectionState,
 } from './AppSectionState';
-
-export interface QueueStatus {
-  totalCount: number;
-  count: number;
-  unknownCount: number;
-  errors: boolean;
-  warnings: boolean;
-  unknownErrors: boolean;
-  unknownWarnings: boolean;
-}
 
 export interface QueueDetailsAppState extends AppSectionState<Queue> {
   params: unknown;
@@ -34,7 +23,6 @@ export interface QueuePagedAppState
 }
 
 interface QueueAppState {
-  status: AppSectionItemState<QueueStatus>;
   details: QueueDetailsAppState;
   paged: QueuePagedAppState;
   options: {
